Add security.

+++ server/webportal/src/main/java/ru/kolaer/server/webportal/mvc/model/dao/DAO.java
@@ -7,6 +7,6 @@ import java.util.List;
  */
 public interface Dao<T> {
     List<T> findAll();
-    T findByID(short id);
+    T findByID(int id);
     void save(T obj);
 }
