+++ server/webportal/src/main/java/ru/kolaer/server/webportal/mvc/model/dao/impl/EmployeeDaoImpl.java
@@ -25,7 +25,7 @@ public class EmployeeDaoImpl implements EmployeeDao {
     }
 
     @Override
-    public GeneralEmployeesEntity findByID(short id) {
+    public GeneralEmployeesEntity findByID(int id) {
         return null;
     }
